refactor(global): pair programming refactor
