+++ src/Global/globalStyles.js
@@ -16,4 +16,6 @@ export const GlobalStyle = createGlobalStyle`
    cursor: pointer;
   }
 
+
+
   `;
