+++ src/Pages/Landing/landingPage.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import LandingComponent from '../../Components/Landing/landing';
+
+export default function LandingPage() {
+    return <LandingComponent />;
+}
