+++ src/Pages/Login/loginPage.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import LoginDialog from '../../Components/LoginDialog/logindialog';
+
+export default function Login() {
+    return <LoginDialog />;
+}
