+++ src/Pages/PostDetails/postDetailsPage.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import Post from '../../Components/Post/post';
+
+export default function PostDetails() {
+    return <Post />;
+}
