+++ src/Services/labEdiit.js
@@ -1,7 +1,7 @@
-import axios from "axios";
+import axios from 'axios';
 
 const labEdiit = axios.create({
-  baseURL: "https://us-central1-labenu-apis.cloudfunctions.net/labEddit/",
+    baseURL: 'https://us-central1-labenu-apis.cloudfunctions.net/labEddit/',
 });
 
 export default labEdiit;
